# PERF: improve performance of nansum for integers